moved login component related files to its own folder. Moved folder to component folder
